trading-bot-possum-pm/requirements.txt
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
pandas>=2.0.0
rapidfuzz>=3.0.0
